test(vhvu-090): relocate 6 chart-component tests into @spaarke/visuals harness
test-diet follow-up (ADR-038 §7 spirit): the BarChart/DonutChart/LineChart/
MetricCard/MiniTable/StatusDistributionBar tests exercised components that moved
to @spaarke/visuals (VHVU-041) but the test files still lived in the PCF and ran
cross-package. Re-home them with the code they test.

- Bootstrap a Jest harness in @spaarke/visuals: jest.config.cjs, babel.config.cjs,
  jest.setup.ts, tsconfig.jest.json (.cjs because the package is "type":"module").
  Add test devDeps (jest/ts-jest/jsdom/babel-jest/@testing-library/{react,dom,
  jest-dom}/@types/jest/scheduler). tsconfig.json now excludes tests from the
  production tsc --noEmit build.
- git mv the 6 test files → src/components/__tests__/; fix imports to '../X'.
- Verified: @spaarke/visuals 71 tests / 6 suites pass; VisualHost 60 tests /
  3 suites (CalendarVisual container + 2 services) pass; tsc --noEmit green.
  Same 131-test coverage, correctly co-located. No runtime source changed →
  VisualHost bundle + deployed v1.4.37 unaffected.
- VisualHost jest React-singleton moduleNameMapper retained (still needed for the
  CalendarVisual container test's cross-package import of the pure component).

Report: projects/visual-host-version-update/notes/test-diet-report.md
