+++ src/client/shared/Spaarke.Visuals/package.json
@@ -14,7 +14,9 @@
   },
   "scripts": {
     "build": "tsc --noEmit",
-    "lint": "tsc --noEmit"
+    "lint": "tsc --noEmit",
+    "test": "jest",
+    "test:watch": "jest --watch"
   },
   "peerDependencies": {
     "@fluentui/react-charting": "^5.0.0",
@@ -24,14 +26,27 @@
     "react-dom": "^16.14.0 || ^17.0.0 || ^18.0.0 || ^19.0.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.28.5",
+    "@babel/preset-env": "^7.28.5",
+    "@babel/preset-react": "^7.28.5",
+    "@babel/preset-typescript": "^7.28.5",
     "@fluentui/react-charting": "^5.23.0",
     "@fluentui/react-components": "^9.46.2",
     "@fluentui/react-icons": "^2.0.200",
+    "@testing-library/dom": "^10.4.1",
+    "@testing-library/jest-dom": "^6.9.1",
+    "@testing-library/react": "^16.3.1",
+    "@types/jest": "^30.0.0",
     "@types/node": "^22.19.19",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
+    "babel-jest": "^30.2.0",
+    "jest": "^30.2.0",
+    "jest-environment-jsdom": "^30.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "scheduler": "^0.27.0",
+    "ts-jest": "^29.4.6",
     "typescript": "^5.3.3"
   }
 }
