+++ src/client/shared/Spaarke.Visuals/src/components/__tests__/MiniTable.test.tsx
@@ -1,11 +1,7 @@
 import * as React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import { FluentProvider, webLightTheme, webDarkTheme } from '@fluentui/react-components';
-import {
-  MiniTable,
-  IMiniTableColumn,
-  IMiniTableItem,
-} from '../../../../../shared/Spaarke.Visuals/src/components/MiniTable';
+import { MiniTable, IMiniTableColumn, IMiniTableItem } from '../MiniTable';
 
 const renderWithTheme = (component: React.ReactElement, theme = webLightTheme) => {
   return render(<FluentProvider theme={theme}>{component}</FluentProvider>);
